requirements.txt
discord.py
easy_pil
python-dotenv